+++ 2022-04-04_test/2022-04-04_test/Fraction.h
@@ -14,9 +14,9 @@ class Fraction
 		Fraction(const Fraction& fraction);
 		~Fraction();
 
-		double getNumerator();
+		long long getNumerator();
 		void setNumerator(double x);
-		double getDenominator();
+		long long getDenominator();
 		void setDenominator(double y);
 
 		Fraction& operator=(const Fraction& fraction);
